--- Content/CHBluePrints/BP_CH4GameMode.uasset
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:262b9f270c9ffb43d24c4e567ac7d7ac55ca7133279f467df80d69b05b0274d9
-size 21697
