--- Content/CHBluePrints/BP_CH4GameStateBase.uasset
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:78ab89c8d02e978cfca1bd3133fbe58278aeed2c291f49a42eff6834981c30be
-size 20866
