--- Content/CHBluePrints/BP_CH4PlayerState.uasset
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:8f2cd469e96a78d32582b08c4e67dfdced2913fb513f532f2da28fe8ddf744b9
-size 20806
