+++ Content/CHBluePrints/Level_BluePrint/BP_CH4PlayerState.uasset
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:bec69386356d50bcd1ed20fb8bf7e50194dc40eb69f029799a63122e1745bbf4
+size 20670
