+++ Source/CH4Project/Public/SpawnVolume/PlayerSpawnVolume.h
@@ -1,0 +1,22 @@
+// Fill out your copyright notice in the Description page of Project Settings.
+
+#pragma once
+
+#include "CoreMinimal.h"
+#include "SpawnVolume/BaseSpawnVolume.h"
+#include "PlayerState/CH4PlayerState.h"
+#include "PlayerSpawnVolume.generated.h"
+
+/**
+ * 
+ */
+UCLASS()
+class CH4PROJECT_API APlayerSpawnVolume : public ABaseSpawnVolume
+{
+	GENERATED_BODY()
+
+public:
+
+	virtual FVector GetSpawnLocation(float Radius = 0.f) const override;
+	
+};
